Compilation fixes and code cleanup
