+++ src/glslUtility.cpp
@@ -2,12 +2,13 @@
 // Written by Varun Sampath, Patrick Cozzi, and Karl Li.
 // Copyright (c) 2012 University of Pennsylvania
 
+#include "glslUtility.hpp"
+
+#include <cstdlib>
+#include <cstring>
 #include <iostream>
 #include <fstream>
 #include <string>
-#include <cstdlib>
-#include <cstring>
-#include "glslUtility.hpp"
 
 using std::ios;
