+++ src/image.cpp
@@ -1,8 +1,9 @@
-#include <iostream>
-#include <string>
+#include "image.h"
+
 #include <stb_image_write.h>
 
-#include "image.h"
+#include <iostream>
+#include <string>
 
 Image::Image(int x, int y)
     : xSize(x), ySize(y), pixels(new glm::vec3[x * y])
