+++ src/interactions.cu
@@ -1,5 +1,9 @@
 #include "interactions.h"
 
+#include "utilities.h"
+
+#include <thrust/random.h>
+
 __host__ __device__ glm::vec3 calculateRandomDirectionInHemisphere(
     glm::vec3 normal,
     thrust::default_random_engine &rng)
