+++ src/scene.h
@@ -1,23 +1,14 @@
 #pragma once
 
-#include <vector>
-#include <sstream>
-#include <fstream>
-#include <iostream>
-#include "glm/glm.hpp"
-#include "utilities.h"
 #include "sceneStructs.h"
-
-using namespace std;
+#include <vector>
 
 class Scene
 {
 private:
-    ifstream fp_in;
     void loadFromJSON(const std::string& jsonName);
 public:
-    Scene(string filename);
-    ~Scene();
+    Scene(std::string filename);
 
     std::vector<Geom> geoms;
     std::vector<Material> materials;
