+++ src/sceneStructs.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <cuda_runtime.h>
+
+#include "glm/glm.hpp"
+
 #include <string>
 #include <vector>
-#include <cuda_runtime.h>
-#include "glm/glm.hpp"
 
 #define BACKGROUND_COLOR (glm::vec3(0.0f))
