+++ src/utilities.cpp
@@ -4,12 +4,13 @@
 //  File: utilities.cpp
 //  A collection/kitchen sink of generally useful functions
 
+#include "utilities.h"
+
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/matrix_inverse.hpp>
-#include <iostream>
-#include <cstdio>
 
-#include "utilities.h"
+#include <cstdio>
+#include <iostream>
 
 float utilityCore::clamp(float f, float min, float max)
 {
